Fix CreateUiContext
