+++ WalletWasabi.Fluent/Models/UI/UiContext.cs
@@ -55,64 +55,4 @@ public class UiContext
 			_navigate?.Navigate(target)
 			?? throw new InvalidOperationException($"{GetType().Name} {nameof(_navigate)} hasn't been initialized.");
 	}
-
-	private static IWalletRepository CreateWalletRepository()
-	{
-		if (Services.WalletManager is { })
-		{
-			return new WalletRepository();
-		}
-		else
-		{
-			return new NullWalletRepository();
-		}
-	}
-
-	private static IHardwareWalletInterface CreateHardwareWalletInterface()
-	{
-		if (Services.WalletManager is { })
-		{
-			return new HardwareWalletInterface();
-		}
-		else
-		{
-			return new NullHardwareWalletInterface();
-		}
-	}
-
-	private static IFileSystem CreateFileSystem()
-	{
-		if (Services.DataDir is { })
-		{
-			return new FileSystemModel();
-		}
-		else
-		{
-			return new NullFileSystem();
-		}
-	}
-
-	private static IClientConfig CreateConfig()
-	{
-		if (Services.PersistentConfig is { })
-		{
-			return new ClientConfigModel();
-		}
-		else
-		{
-			return new NullClientConfig();
-		}
-	}
-
-	private static IApplicationSettings CreateApplicationSettings()
-	{
-		if (Services.PersistentConfig is { } persistentConfig && Services.UiConfig is { } uiConfig)
-		{
-			return new ApplicationSettings(persistentConfig, uiConfig);
-		}
-		else
-		{
-			return new NullApplicationSettings();
-		}
-	}
 }
